backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
pyjwt
sendgrid>=6.10.0
email-validator
python-multipart
